+++ arcs/src/components/drawing_object.rs
@@ -1,4 +1,7 @@
-use crate::{Arc, Line, Point};
+use crate::{
+    algorithms::{Bounded, Closest, ClosestPoint, Translate},
+    Arc, BoundingBox, Line, Point, Vector,
+};
 use specs::prelude::*;
 
 // for rustdoc links
@@ -24,4 +27,46 @@ pub enum Geometry {
     Line(Line),
     Arc(Arc),
     Point(Point),
+}
+
+impl ClosestPoint for Geometry {
+    fn closest_point(&self, target: Point) -> Closest {
+        match self {
+            Geometry::Point(p) => p.closest_point(target),
+            Geometry::Line(l) => l.closest_point(target),
+            Geometry::Arc(a) => a.closest_point(target),
+        }
+    }
+}
+
+impl ClosestPoint for DrawingObject {
+    fn closest_point(&self, target: Point) -> Closest {
+        self.geometry.closest_point(target)
+    }
+}
+
+impl Bounded for Geometry {
+    fn bounding_box(&self) -> BoundingBox {
+        match self {
+            Geometry::Line(line) => line.bounding_box(),
+            Geometry::Arc(arc) => arc.bounding_box(),
+            Geometry::Point(point) => point.bounding_box(),
+        }
+    }
+}
+
+impl Translate for Geometry {
+    fn translate(&mut self, displacement: Vector) {
+        match self {
+            Geometry::Point(ref mut point) => point.translate(displacement),
+            Geometry::Line(ref mut line) => line.translate(displacement),
+            Geometry::Arc(ref mut arc) => arc.translate(displacement),
+        }
+    }
+}
+
+impl Translate for DrawingObject {
+    fn translate(&mut self, displacement: Vector) {
+        self.geometry.translate(displacement);
+    }
 }
