+++ arcs/src/components/viewport.rs
@@ -1,4 +1,4 @@
-use crate::{CanvasSpace, DrawingSpace, Point};
+use crate::{algorithms::Translate, CanvasSpace, DrawingSpace, Point, Vector};
 use euclid::Scale;
 use specs::prelude::*;
 use specs_derive::Component;
@@ -10,4 +10,20 @@ pub struct Viewport {
     pub centre: Point,
     /// The number of pixels each drawing unit should take up on the screen.
     pub pixels_per_drawing_unit: Scale<f64, DrawingSpace, CanvasSpace>,
+}
+
+impl crate::algorithms::Scale for Viewport {
+    /// Zoom the viewport, where a positive `scale_factor` will zoom in.
+    fn scale(&mut self, scale_factor: f64) {
+        assert!(scale_factor.is_finite() && scale_factor != 0.0);
+        self.pixels_per_drawing_unit = euclid::Scale::new(
+            self.pixels_per_drawing_unit.get() / scale_factor,
+        );
+    }
+}
+
+impl Translate for Viewport {
+    fn translate(&mut self, displacement: Vector) {
+        self.centre.translate(displacement);
+    }
 }
