--- arcs/src/primitives/mod.rs
@@ -1,7 +1,0 @@
-//! Basic geometric types which are generic over their coordinate space.
-
-mod arc;
-mod line;
-
-pub use arc::Arc;
-pub use line::Line;
