+++ core/Cargo.toml
@@ -8,3 +8,8 @@ edition = "2018"
 
 [dependencies]
 euclid = "0.20.11"
+specs = { version = "0.16.1", optional = true }
+
+[features]
+default = []
+ecs = ["specs"]
